UI Improvements (#136)
* Only display tiles for model runs that are selected

The previous default behavior: when no model runs are selected in
the UI, render *all* model runs' tiles on the map.
New behavior: only render the currently selected model runs' tiles on
the map. If none are selected, no tiles are rendered.

* Remove ground truth toggle

* Don't move camera when filtering by region

* Remove console.log statement

* Add toggle for showing site outlines

Now, site outlines are off by default and must be toggled on to be seen.