Don't calculate bounding box across all model runs if no region is selected
This attempts to limit the amount of polygons that we're asking PostGIS
to dynamically calculate a bounding box for in the model runs listing
endpoint. If the user doesn't supply a region to filter by, the server
will now *not* send back a bounding box to avoid an expensive PostGIS
query across every polygon in the system.