Fix errors in site model ingest (#165)
* Fix bug in pydantic model validators

When validating a field in a pydantic model, it's not safe to assume
that other fields are defined yet.

* Use pydantic descriminator for improved ingest error messages

* Cleaner support for comma-seperated strings in site model

* Don't support out-of-spec format for region models
Only site model `Observations` can use comma-seperated strings as lists
of certain fields; this is not supported by the region spec.

* Fix incorrect ingest/data model for site observations

* Add test for site model ingest

* Fix response code for ingest endpoints

Previously, these endpoints we (incorrectly) returning 200 responses
upon successful ingest instead of 201. This commit tweaks the
django-ninja route configs to properly return a 201, plus the IDs of the
newly ingested SiteEvaluations.

* Make django-ninja TestClient session-scoped

Apparently it needs to be a singleton, and will throw an error if
you try to instantiate it more than once in a test run.