Initial T&E Scoring Database Connection (#299)
* WIP T&E connection

Co-authored-by: Wang, Sean S <sean.wang@jhuapl.edu>
Co-authored-by: JHUAPL <>

* Restructure scoring `api.py` to get swagger docs working

This makes it so the swagger endpoint docs render correctly at
`/api/scoring/docs`.

* Misc fixes to get endpoints working properly

* Fix linting

* Fix potential 500 error

If `.get()` is called on a non-existant object, it will raise an
exception and result in a 500 error. The correct way to query for
an object in this case is to use `get_object_or_404`.

* Replace `.annotate` with `.alias` where possible

* Import response schemas from rdwatch instead of redefining

* Synchronize scoring models and migrations

Some scoring DB models didn't have an associated `CreateModel`
migration, while others had a migration but were not encoded as models,
causing the `check-migrations` CI check to fail. This commit adds
migrations for models without one, and removes migrations for tables
that are not (yet?) encoded as Django models.

* Apply suggestions from code review

Co-authored-by: Bryon Lewis <61746913+BryonLewis@users.noreply.github.com>

* Add missing import

* Revert removing area calculation

---------

Co-authored-by: Wang, Sean S <sean.wang@jhuapl.edu>
Co-authored-by: Bryon Lewis <61746913+BryonLewis@users.noreply.github.com>