Optimize sites list query
Moves the color_code subquery to a separate query to avoid a slow
subquery in the main query, due to the
EvaluationBroadAreaSearchDetection table being extremely large.
The number of sites in an evaluation run is small enough that it's
feasible to load the site_id->color_code mapping into memory and
manually join it with the sites_list in Python.